Move building related objects to the new `buildings` app
