+++ seismic_site/buildings/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class BuildingsConfig(AppConfig):
+    name = 'buildings'
