+++ seismic_site/seismic_site/settings.py
@@ -42,7 +42,8 @@ INSTALLED_APPS = [
     'seismic_site.apps.SeismicSiteConfig',
     'api.apps.ApiConfig',
     'map_app.apps.MapAppConfig',
-    'crispy_forms'
+    'crispy_forms',
+    'buildings'
 ]
 
 MIDDLEWARE = [
